requirements.txt
uvicorn
fastapi
boto3
python-multipart
python-dotenv
pymongo[srv]
fastapi-sso